Added tests for PacketSerializer and MessageSerializer
Former-commit-id: 940cc3e97e5a606c4afe3763c030db7086c9ab80

+++ fingerprintscanner/src/main/java/com/simprints/fingerprintscanner/v2/domain/message/un20/Un20MessageProtocol.kt
@@ -27,7 +27,7 @@ object Un20MessageProtocol: MessageProtocol {
 
     fun buildMessageBytes(un20MessageType: Un20MessageType, data: ByteArray): ByteArray {
         val length = data.size
-        val header = un20MessageType.getBytes() + length.toShort().toByteArray()
+        val header = un20MessageType.getBytes() + length.toByteArray()
         return header + data
     }
 }
